Add non-optimal solution of Subset Sum
On lookup query search we need look in time complexity O(log(n)).
Current solution is O(n), not good enough.

The next strategy is prepare the data with the max-sums for each
subset as nodes of a binary search tree. That way we can do later
a algorithm to check the smallest subset based on that date with
complexity of O(log(n)).